requirements.txt
deepface>=0.0.93
tensorflow>=2.15.0
fastapi>=0.109.0
uvicorn>=0.27.0
click>=8.1.0
pillow>=10.0.0
pillow-heif>=0.15.0
pyyaml>=6.0.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
